Database and ConnectionPool can now execute connection preparation closures.
Being able to execute pragma statements prior to running additional statements on a connection is very useful, but is still too limiting. Instead, you need to have full freedom to prepare a connection in any way you see fit. For example, you may need to register a custom collation closure with all connections before running a query. There are too many possible scenarios to support, so the easiest way is to add a preparation closure giving the user the freedom to prepare the connection in any way they see fit.